refactor: 一つの処理が大きくなりすぎたので、修正。
